Port of the *module-provider-functions* extension to cl:require from SBCL.
The old defmodule handler and the module:*-library.* loading have been
factored out to their own module-provider functions, which are by default
on *module-provider-functions*.  Note however that this implies one
functional change for defmodule defined modules:  The loading of the files
constituting such a module are now _NOT_ wrapped with without-package-locks.

Wrapping of module:*-library.* files with without-package-locks is only
retained temporarily, until proper changes to the various modules have
been made.

Also adjusted documentation of require to correctly state the default for
*require-verbose* which has always been t, not nil as previously claimed.

This change is still missing documentation in the user manual.