Commit of the compiler backend portion of the PPC/Darwin port.
This includes various fixes from SBCL and other sources to the PPC backend.
It also does away with Gary Byers' renumbering of the type tags for
function and instance pointers, which allowed for a slightly clever and
presumably faster function calling convention.  These changes have been
conditionalized out with the PPC-FUN-HACK conditional, so that we might
revisit the changes in the future.  The current port has only been tested
without PPC-FUN-HACK, but this might change.