Fix a bunch of FORMAT bugs caught by ansi-tests:
o "~,,,',A" was incorrectly parsed:  The fill character is ",".
o (FORMAT NIL "~VD" NIL 7) was not treating a NIL value for V to mean
  use the default.
o "~3,,-1A" would add one pad character.  I think this is allowed, but
  not the intent.  The extra pad character(s) are no longer produced.
o ~F was printing numbers in exponential notation if the number was
  too large or small.  We know print out all the necessary digits,
  even though we're allowed to use exponential notation if more than
  100 digits would be produced.
o Correctly handle justification (~< ~>) with non-zero minpad value.