COMPUTE-TAB-SIZE: o Previous changes were wrong.  This version handles relative and   section-relative columns correctly,  I think.
PPRINT-TABULAR
o The computation of colinc was wrong---I don't think we need to add
  one to tabsize.
o I don't think we should output the space character.  It results in
  one extra space.

With these changes, the format tests in ansi-tests for ~T and friends
pass.  I think these are right now, but I'm not 100% sure.