Implement tracing of flet/labels functions.  This probably needs more work and could probably be implemented better.
With these changes (trace (labels foo bar)) will trace the labels
function FOO in the function BAR.  We only support encapsulate nil,
here.  No check is made for this.

code/ntrace.lisp:
o In TRACE-FDEFINITION, recognize a list as a valid function, and
  return the list as the value of TRACE-FDEFINITION.  This seems
  wrong, but I'm not sure if there's a real fdefinition for it, or if
  we could create a fake one.

code/debug-int.lisp:
o In FUNCTION-DEBUG-FUNCTION, recognize a list as the name of a
  function, and find the corresponding compiled-debug-function and
  create and return the new compiled-debug-function.