Yet another version of INTEGER-DECODE-DOUBLE-DOUBLE-FLOAT.  The previous version was wrong for the double-double (2d0, -4d17) which caused printing to be off by a factor of two.
I think this version is right, but the number of bits in the integer
part can be more than 106 (double-float-digits), because double-double
floats can have more than 106 bits in them.  There's a problem with
converting the result back to a double-doube---the result isn't
exactly the same.  I think this is a bug in DOUBLE-DOUBLE-FROM-BITS in
bignum.lisp.