Was not getting the right instruction for the faulting floating-point exception.
This was happening when the faulting FP instruction was in the delay
slot of a branch.  The PC was incremented to the branch target, and
hence had the wrong information.  The FPQ structure has the right
information.

code/sparc-svr4-vm.lisp:
o Add FPQ structure so we can access the information about the
  faulting FP instruction.

compiler/sparc/float.lisp:
o Extract the correct information about the faulting FP instruction
  instead of from the PC.