Clear the FPU status word before loading the control word from the sigcontext.
This is a workaround for an issue on x86/darwin where loading of the
control word enabling the invalid operation exception would cause that
exception because the status word had that exception.  Not sure why
this is needed or why the invalid operation exception was set.  It was
happening in a call run-program while concatenating the files for the
hemlock library.