Fix minor issue:
(defun zot (x)
  (declare (type (INTEGER -536871936 536871934) x)
	   (optimize (speed 3)))
  (kernel::%double-float x))

produces a compiler note being unable to optimize due to type
uncertainty x.  This is confusing because the vop does get used, and
we get a nice single intruction.  This is caused by the unsigned-byte
32 deftransform for %single-float/%double-float

Work around the issue by

o adding new deftransform to convert %single-float/%double-float to
  %%single-float/%%double-float,
o adjusting the deftransform for unsigned 32-bit to call %%foo,
o adding a deftransform to always convert %foo to %%foo for signed
  32-bit arg,
o changing the vop translation to %%foo instead of %foo.