Fix long-standing issue on sparc and ppc with linkage-tables.
If the C code changes, genesis was storing undefined_tramp and
closure_tramp with the addresses of the lisp used to build.  But that
wouldn't match the values of the lisp used to run, which would cause
weird segfaults and core dumps.

Hence, we need to load up lisp.nm so we can get the correct values for
undefined_tramp and closure_tramp.