Speed up bignum-truncate by getting rid of the special variables *truncate-x* and *truncate-y*.  Accessing these in tight loops hurts performance.  So pass truncate-x and truncate-y as parameters to the routines that need them.  Reorder the truncate support routines so we can block-compile all of the routines for truncate.
With these changes, we see the following changes according to
cl-bench.  The first column and reference is 2008-10 snapshot
(darwin), and the next column is the ratio of this new code to the
reference.  So we see improvements from 10% to 50%.

BIGNUM/ELEM-100-1000     [      0.23]   0.84
BIGNUM/ELEM-1000-100     [      0.50]   0.89
BIGNUM/ELEM-10000-1      [      0.51]   0.90
BIGNUM/PARI-100-10       [      0.05]   0.63
BIGNUM/PARI-200-5        [      0.22]   0.51
PI-DECIMAL/SMALL         [      2.44]   0.70
PI-DECIMAL/BIG           [      2.24]   0.73
PI-ATAN                  [      1.07]   1.00
PI-RATIOS                [      2.56]   0.88