Reduce the size of the dispatch-table in the readtable.  Previously, we used a huge array of 65536 elements.  Now use a structure containing a 256-element vector and a hash-table.
o Define a new structure (for unicode) for the dispatch table, as
  described above.
o Abstract out the access to the dispatch table by adding the
  functions GET-DISPATCH-CHAR, SET-DISPATCH-CHAR, and
  COPY-CHAR-DISPATCH-TABLE.
o Use the new functions to access the dispatch table.

(The non-unicode stuff is implemented, but currently untested as of
this checkin.)

Simple tests using ansi-tests from gcl indicate that dispatch-table
had a fairly large impact on the speed of the tests.  This change,
combined with the previous change for the character attribute and
macro tables makes the ansi-tests run somewhat slower than before
instead of 5-10 times slower.