Add support for Unicode NFC and NFKC forms.  Implement STRING-TO-NFC and STRING-TO-NFKC.
This probably needs some more work.  The composition table should
probably be a trie and should be in unidata.bin instead of the hash
table that we use now.  The composition exclusion list should be
probably be in unidata.bin too instead of here.

These functions pass all of the normalization tests.