Merge changes from sparc-tramp-assem-2010-07-19.  This moves closure_tramp and undefined_tramp from the C side to the Lisp side via Lisp assembly routines.  (Simplifies things quite a bit for sparc.)
Also includes some GC fixes for sparc and ppc to handle moving the pc,
npc, lr, and ctr registers.  They need to be handled like the lip
register.

See bootfiles/20a/boot-2010-07-1-cross.lisp for instructions on how to
bootstrap this change.  A basic cross-compile with a specific cross
bootstrap file is needed.