Fix Trac #43:  unread-char doesn't change file-position
The issue is caused by FAST-READ-CHAR-STRING-REFILL.  In some
situations the number of octets converted is not exactly the same as
the length of the buffer because the last octets in the buffer don't
hold a complete encoded character.  This is ok, but we didn't update
the ibuf-head pointer to indicate that some octets haven't actually
been converted.  This confuses FILE-POSITION.