Fix ticket:74
When accounting for the octets left in the in-buffer that we haven't
read (or converted to characters), we were subtracting the index from
the total in-buffer length.  This is wrong if the file is less than
the total in-buffer length.  We should have subtracted from the actual
number of octets in the in-buffer.