Fix some issues with zero check in gc_free_heap.
In gc_free_heap, the check that unallocated pages are zero can only be
done if gencgc_unmap_zero is MODE_MAP or MODE_MEMSET.  The other modes
are known not to zero memory.

Also, don't limit the check to the first 16K pages, and also check the
entire page of GC_PAGE_SIZE bytes, not just the first 1024 words.