Fix many issues and implement a few more missing pieces.
The fixes:
o Forgot to define a printer for the reg MOV case.
o Use the correct values for the emitter for the register MOV inst.
o Comment out the flex-operand case. We don't support that.
o The split-imm16-printer was not printing the imm16 value out
  correctly. 
o Add some comments to the format-2-immed-printer to make it clearer
  what's happening.  (We had the U and W values backwards.)
o Rename DECODE-INDEXING-MODE to DECODE-IMMEDIATE-INDEXING-MODE to
  make it clearer.  Add some comments too.
o Add comments to DECODE-INDEXING-AND-OPTIONS.
o Fix DEFINE-LOAD/STORE-INST.
  o We were not generating the correct instruction for stores.
  o Need to compute the reg encoding for the indexing-mode register.
o Better implementation of NOT-IMPLEMENTED that takes an arg saves
  that to r0 so the trap handler can figure out what VOP is not
  implemented.
 
The new stuff:
o Implement EMIT-COMPUTE-INST better, and always use the worst-case
  for now.