Produce correct results on branch cuts for atanh.
For (atanh x), where x is a real number on the branch cuts for atanh,
make the code return the correct value based on the definition of
atanh. This is inconsistent with the description in the CLHS because
the values on the branch cut are now continuous with different
quadrants.  The formula is clear, however, for the values on the
branch cut.

 * src/code/irrat.lisp
 * src/code/irrat-dd.lisp
   * Use the correct branch cut values for atanh.
 * tests/trig.lisp
   * Update tests accordingly, with additional comments.