Fix bug in DOUBLE-FLOAT-BITS vop.
Add a lifetime spec to the arg so that we don't accidentally destroy
the arg when writing the two results.  Previously, HI-BITS was
allocated to the same register as FLOAT (in a descriptor). Writing
HI-BITS would destroy the contents of the FLOAT and when trying to
write lo-bits, we would read from the wrong address because FLOAT now
contains the high bits of the double float number.