Re-enable the x87 floating-point mode stuff.
On 32-bit linux, we can still get FP exceptions using x87 because
32-bit linux can still use x87 instructions for arithmetic.  Because
of this, we need to re-enable the support x87 floating-point modes,
including getting and setting the modes and also extracting the modes
from a sigcontext.

 * src/code/float-trap.lisp:
   * Put back support for getting and setting the x87 FP modes. 
 * src/compiler/x86/float.lisp:
   * Add comment on the layout of the status and control words for
     x87. 
 * src/lisp/Linux-os.c:
   * Put back support for getting the x87 (and sse2) FP modes. Needed
     in the sigfpe-handler in float-trap.lisp.

Some of this needs to be cleaned up because we always require sse2
now.