Fix issue with negative value for lisp::cycles-per-tick.
This causes negative cpu cyles with TIME. Basic issue is that a cpu
frequency of 2.3 GHZ won't fit in an int. Use an unsigned int.

We also take this opportunity to use a rounded value for
clocks-per-tick instead of truncating.  For this particular case the
ratio is actual 68.99 which would truncated to 68. We should probably
use 69 instead.