Remove darwin-lispadjuster.c and ppc-darwin-dlshim.c
Use of ppc-darwin-dlshim.c was removed, but I forgot to remove the
file.

Use of darwin-lispadjuster.c has been commented out for a while, so
I'm getting rid of it. It seems at one point it was needed to resize
the CMUCLRO section in the lisp executable, but there is CMUCLRO
section anymore.

darwin-lispadjuster.c:
 * Deleted

ppc-darwin-dlshim.c:

Config.ppc_darwin:
 * Remove Make rule for adjustlisp and friends.
 * Remove old commented out version of OS_SRC; we don't use
   ppc-darwin-dlshim.c anymore.