Initialize fpu_type to SSE2 instead of AUTO.
This shows up when creating executables.  Executables don't have the
magic number in the core file to indicate whether the core suports
sse2 or x87, so the fpu_type is never set to the actual type, and
fpu_mode is set to the value of fpu_type.  But there are asserts in
the C code to check that fpu_mode is SSE2 since we have dropped
support for x87. 

One case where this happens is building maxima with a cmucl executable
on darwin.  An assert triggers preventing maxima from running.  This
change fixes that.