Reimplement NOT-IMPLEMENTED to dump the symbol name into the code stream.
See issue #41.  The symbol name is now saved into the instruction
stream as a sequence of bytes.  The length is encoded into a branch
instruction before the bytes so that if the trap is continued, the
bytes are skipped over nicely.

Had to update how EMIT-NOT-IMPLEMENTED was calling NOT-IMPLEMENTED and
a few other places where bare NOT-IMPLEMENTED was called directly.