Fix data-vector-set-c for unsigned-byte 1, 2, and 4
For simple-arrays of 1, 2, or 4-bit elements, data-vector-set-c was
incorrectly merging the new value into the array when the index is a
multiple of the number of elements per (32-bit) word.  Thus, for 4-bit
elements, the new value was not merged in when the index is a multiple
of 8.  In these cases, there's no need to shift the array value or the
new value to move them into the correct place.  When the shift is
zero, the code accidentally removed the part that merges in the new
value.

Fix #10.