Fix #34: Handle newline character correctly
The string containing #\Newline that is used to output the newline
character worked on little-endian architectures because the low octet
was first in memory.  However on a big-endian sparc, a NUL character
was output instead since the high octet is first in memory.

So, create a explicit unsigned-byte 8 array containing exactly 1 octet
that represents a newline character code.

Tests pass on both sparc and x86/linux.