Replace version arch os args with switches
The positional version, arch, and os args are now commandline
switches.  This allows the user to specify arch and os names, for
example, without specifying a version.  This allows the version to be
defaulted.  Previously, the version had to be specified if arch or os
were desired.  

The original default values are used if these options aren't given.