Don't flame out when changing constants
Stolen from cross-x86-amd64 so that defconstant doesn't flame out when
changing the value of a defconstant.

Also, don't frob old-vm:word-bits with a new value.  Not sure this is
needed, but it gets rid of a problem when compiling. (Why does the
old-vm stuff leak into the new vm anyway?)