Use correct offset when printing lispobj function address
The previous version had a hardwired #x17 (23 dec) to get the object
address from the function code offset address.  Compute this value
from function-code-offset instead so that it will work for 64-bit
lisps too.

Also increase the size of the field to 12 hex digits.  That should be
enough bits (48 bits) of address space for any thing we're likely to
run on.