Print out correct info for sparc (and ppc)
The values for the initially defined functions in lisp.map were
incorrect for sparc and ppc where the raw-addr slot of function is not
the address of the code.  It is has function-pointer-type so the code
actually starts at the function-code-offset.  So for sparc and ppc,
continue to print out the raw-addr slot, but print out the actual code
start too (instead of some number preceding the function object).

Add comment in the code for this too.