Pad data block sizes to 16-byte boundary.
Data blocks are padded to double-word boundaries which is 16 bytes for
sparc64.

This also means telling new-genesis round up by 16 bytes in
allocate-descriptor.

Adjust string-to-core for 64-bit objects too.

At this point, the static symbols appear to print out correctly (as
determined by call print (in the lisp monitor).  Some objects, such as
*fp-constant-0d0* however appear to be incorrect.  The value pointer
appears to be a double-float but the object is messed up.  But maybe
this is caused by a buggy print function.